flat_mutation_reader: add slicing_forwarding_reader
Implements slicing on a reader, using fast-forwarding. In other words:
* given an underlying reader created with schema::full_slice()
* a partition-slice with row-ranges

This reader will fast forward the underlying reader to each individual
row-range in the slice, such that the output is the same as if said
reader was created with the slice in the first place.
Useful for when a reader is reused across many partitions, having to
read a different slice from each one.