db/view/view_updating_consumer: reuse baseline reader across partitions
Currently the baseline reader -- the reader which provides the base
against which the diff with the new writes is calculated -- is created
and thrown away after each partition or chunk of partition the view
updating consumer consumes. This is very inefficient as with many small
partitions the cost of creating/destroying the reader can add very
significant overhead. As we are iterating over partitions in their
natural order, we can reuse a reader across all the partitions,
fast-forwarding it to the current partition.