db/view/view_update_generator: process staging sstables in small batches
Currently in each iteration of the view update generator loop, all
accumulated staging sstables of all tables are processed, one table at a
time. If for any reason staging sstables manage to pile up, this results
in a "cliff" behaviour. All staging sstables for a table are processed
at once and are all moved to the main sstable group only after all of
them are processed. If for any reason the processing fails in the
middle, the whole thing has to be restarted, with no incremental
progress possible.
This patch changes the loop so it processes sstables in small batches of
4. One iteration processes a batch from all tables, one table at a time.
This makes process incremental. Even if a lot of staging sstable files
managed to accumulate, a small batch has a much better chance of
finishing without problems, leading to improved resiliency as if there
is a failure, only the last batch has to be retried.
As staging sstables are assumed to be disjoint, processing them in small
batches doesn't create any duplicate work for us: we don't expect any
partition to be found in more than a single sstable.

Fixes: #11744